move spotlight pokemon to show for all users
